+++ apps/spark/bin/set-version.js
@@ -2,15 +2,17 @@
 
 "use strict";
 
-const version = require("../../../package.json").version;
 const fs = require("fs");
 const path = require("path");
+const version = require("../../../package.json").version;
+
+require('dotenv').config();
 
 const versionFile = {
-  "version": version
+  "version": version + ((process.env.VITE_CAMPAIGN_ENABLED === "true") ? " with campaigns" : "")
 }
 
-fs.mkdir(path.join(__dirname, "../src/__generated__"), { recursive: true}, (err) => {
+fs.mkdir(path.join(__dirname, "../src/__generated__"), { recursive: true}, () => {
   fs.writeFile(path.join(__dirname, "../src/__generated__/version.json"), JSON.stringify(versionFile), err => {
     if (err) {
       console.error("Error writing version.json", err);