+++ apps/spark/src/pages/ProductPage.tsx
@@ -1,6 +1,13 @@
 import React, { FC } from "react";
 import { match } from "react-router-dom";
-import { ProductImpl, useProductByIdQuery } from "@coremedia-labs/graphql-layer";
+import {
+  ProductByIdWithCampaignsQuery,
+  ProductImpl,
+  Slot,
+  SlotResult,
+  useProductByIdQuery,
+  useProductByIdWithCampaignsQuery,
+} from "@coremedia-labs/graphql-layer";
 import Loading from "../components/Loading/Loading";
 import { ApolloClientAlert, ProductNotFoundAlert } from "../components/Error/Alert";
 import { DetailProduct } from "../models/Detail/DetailProduct";
@@ -12,6 +19,13 @@ import { initializeProductBannerFromProduct } from "../models/Banner/ProductBann
 import { useSiteContextState } from "../context/SiteContextProvider";
 import DetailedProduct from "../components/Product/DetailedProduct";
 import ProductPageContext from "../context/ProductPageContext";
+import {
+  CAMPAIGN_CONTEXT_PRODUCT,
+  getRefinementData,
+  hasCampaignData,
+  isCampaignEnabled,
+} from "../utils/Campaign/CampaignUtil";
+import { Download } from "../models/Detail/DetailCMProduct";
 
 interface PageProps {
   match: match<RouteProps>;
@@ -23,13 +37,22 @@ interface RouteProps {
 }
 
 const ProductPage: FC<PageProps> = ({ match }) => {
-  const { siteId, rootSegment } = useSiteContextState();
-  const { data, loading, error } = useProductByIdQuery({
+  const { siteId, currentNavigation, rootSegment } = useSiteContextState();
+  let campaignDataSlots: Array<Slot> | undefined;
+
+  // Fetch page data with or without campaigns
+  const queryVars = {
     variables: {
       externalId: match.params.seoSegment,
       siteId: siteId,
+      refinements: getRefinementData(currentNavigation, CAMPAIGN_CONTEXT_PRODUCT),
     },
-  });
+  };
+
+  const { data, loading, error } = isCampaignEnabled()
+    ? useProductByIdWithCampaignsQuery(queryVars)
+    : useProductByIdQuery(queryVars);
+
   if (loading) {
     return <Loading />;
   }
@@ -56,6 +79,13 @@ const ProductPage: FC<PageProps> = ({ match }) => {
     });
   }
 
+  let downloads: Array<Download> = [];
+  if (product.augmentation && product.augmentation.downloads && product.augmentation.downloads.length > 0) {
+    downloads = product.augmentation?.downloads.map((item: any) => {
+      return item;
+    });
+  }
+
   const detailProduct: DetailProduct = {
     ...initializeProductBannerFromProduct(product as ProductImpl, rootSegment),
     id: product.shortId,
@@ -64,14 +94,23 @@ const ProductPage: FC<PageProps> = ({ match }) => {
     longDescription: product.longDescription,
     shopNowConfiguration: true,
     pictures: media,
+    downloads: downloads,
   };
   const placements = product?.augmentation?.pdpPagegrid?.placements as Placements;
 
+  if (isCampaignEnabled()) {
+    const campaignData = (data as ProductByIdWithCampaignsQuery)?.campaignContent as SlotResult;
+    if (hasCampaignData(campaignData)) {
+      console.debug("Campaign data loaded: ", campaignData);
+      campaignDataSlots = campaignData.slots as Array<Slot>;
+    }
+  }
+
   return (
-    <ProductPageContext media={media} product={detailProduct}>
+    <ProductPageContext media={media} downloads={downloads} product={detailProduct}>
       <SeoHeader title={detailProduct.name} />
       <RootPreviewId metadataRoot={detailProduct.metadata?.root} />
-      <DetailedProduct placements={placements} />
+      <DetailedProduct placements={placements} campaignDataSlots={campaignDataSlots} />
     </ProductPageContext>
   );
 };